+++ src/main/java/com/github/mnesikos/simplycats/entity/goal/CatAttackGoal.java
@@ -50,7 +50,7 @@ public class CatAttackGoal extends Goal {
     @Override
     public void tick() {
         this.cat.getLookControl().setLookAt(this.target, 30.0F, 30.0F);
-        double d0 = (double) (this.cat.getBbWidth() * 2.0F * this.cat.getBbWidth() * 2.0F);
+        double d0 = this.cat.getBbWidth() * 2.0F * this.cat.getBbWidth() * 2.0F;
         double distanceSq = this.cat.distanceToSqr(this.target.getX(), this.target.getY(), this.target.getZ());
         double speed = 0.8D;